+++ oppia/api/resources.py
@@ -10,6 +10,7 @@ from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login
 from django.conf import settings
 from django.core import serializers
+from django.core.mail import send_mail
 from django.core.servers.basehttp import FileWrapper
 from django.db import IntegrityError
 from django.http import HttpRequest, HttpResponse ,Http404
@@ -220,12 +221,13 @@ class ResetPasswordResource(ModelResource):
     ''' 
     For resetting user password
     '''
+    message = fields.CharField()
     
     class Meta:
         queryset = User.objects.all()
         resource_name = 'reset'
         allowed_methods = ['post']
-        fields = ['username']
+        fields = ['username','message']
         authorization = Authorization() 
         always_return_data = True 
         include_resource_uri = False   
@@ -237,63 +239,31 @@ class ResetPasswordResource(ModelResource):
                 bundle.data[r]
             except KeyError:
                 raise BadRequest(_(u'Please enter your %s') % r)
-            
-        '''    
-        data = {'username': bundle.data['username'],
-                'password': bundle.data['password'],
-                'password_again': bundle.data['passwordagain'],
-                'email': bundle.data['email'],
-                'first_name': bundle.data['firstname'],
-                'last_name': bundle.data['lastname'],}
-        rf = RegisterForm(data)
-        if not rf.is_valid():
-            str = ""
-            for key, value in rf.errors.items():
-                for error in value:
-                    str += error + "\n"
-            raise BadRequest(str)
-        else:
-            username = bundle.data['username']
-            password = bundle.data['password']
-            email = bundle.data['email']
-            first_name = bundle.data['firstname']
-            last_name = bundle.data['lastname']
+         
+        bundle.obj.username = bundle.data['username']
         try:
-            bundle.obj = User.objects.create_user(username, email, password)
-            bundle.obj.first_name = first_name
-            bundle.obj.last_name = last_name
-            bundle.obj.save()
-            
-            user_profile = UserProfile()
-            user_profile.user = bundle.obj
-            if 'jobtitle' in bundle.data:
-                user_profile.job_title = bundle.data['jobtitle']
-            if 'organisation' in bundle.data:
-                user_profile.organisation = bundle.data['organisation']
-            user_profile.save()
-            
-            u = authenticate(username=username, password=password)
-            if u is not None:
-                if u.is_active:
-                    login(bundle.request, u)
-                    # Add to tracker
-                    tracker = Tracker()
-                    tracker.user = u
-                    tracker.type = 'register'
-                    tracker.ip = bundle.request.META.get('REMOTE_ADDR','0.0.0.0')
-                    tracker.agent =bundle.request.META.get('HTTP_USER_AGENT','unknown')
-                    tracker.save()
-            key = ApiKey.objects.get(user = u)
-            bundle.data['api_key'] = key.key
-        except IntegrityError:
-            raise BadRequest(_(u'Username "%s" already in use, please select another' % username))
-        del bundle.data['passwordagain']
-        del bundle.data['password']
-        del bundle.data['firstname']
-        del bundle.data['lastname']
-        '''
+            user = User.objects.get(username__exact=bundle.obj.username)
+            newpass = User.objects.make_random_password(length=8)
+            user.set_password(newpass)
+            user.save()
+            if bundle.request.is_secure():
+                prefix = 'https://'
+            else:
+                prefix = 'http://'
+            # TODO - better way to manage email message content
+            send_mail('OppiaMobile: Password reset', 'Here is your new password for OppiaMobile: '+newpass 
+                      + '\n\nWhen you next log in you can update your password to something more memorable.' 
+                      + '\n\n' + prefix + bundle.request.META['SERVER_NAME'] , 
+                      settings.SERVER_EMAIL, [user.email], fail_silently=False) 
+        except User.DoesNotExist:
+            pass
+        
         return bundle       
         
+    def dehydrate_message(self,bundle):
+        message = _(u'An email has been sent to your registered email address with your new password')
+        return message
+    
 class TrackerResource(ModelResource):
     ''' 
     Submitting a Tracker