+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Change Log
 v0.4.1 - released: XXXXX
 ----------------------------
 * https://github.com/DigitalCampus/oppia-mobile-android/issues/283 - add optional job title and organisation to profile
+* Reset password facility https://github.com/DigitalCampus/oppia-mobile-android/issues/issue/292
 
 v0.4.0 - released: 1-Apr-2014
 ------------------------------